Conflicts:

	src/imageSource

+++ build_clean.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+echo -n 'Are you sure? '; read YN
+
+if [ "$YN" != "Y" ]; then
+    exit 1
+fi
+
 git clean -fd &&
     git reset --hard &&
     autoreconf -vi &&
